enable tests
