+++ features/admin/user/managing_administrators/changing_administrator_avatar.feature
@@ -7,7 +7,7 @@ Feature: Changing an administrator's avatar
     Background:
         Given I am logged in as an administrator
 
-    @ui @api @avatar # temporarily enabled to test only the avatar
+    @ui @api
     Scenario: Changing an avatar of an administrator
         Given I have the "ford.jpg" image as my avatar
         And I am editing my details
