+++ features/admin/user/managing_administrators/removing_administrator_avatar.feature
@@ -8,7 +8,7 @@ Feature: Removing an administrator avatar
         Given I am logged in as an administrator
         And this administrator has the "troll.jpg" image as avatar
 
-    @ui @api @avatar # temporarily enabled to test only the avatar
+    @ui @api
     Scenario: Removing an administrator avatar
         Given I am editing my details
         When I remove the avatar
