+++ src/Sylius/Behat/Resources/config/suites/api/user/managing_administrators.yml
@@ -21,5 +21,5 @@ default:
                 - sylius.behat.context.api.admin.save
 
             filters:
-                tags: "@managing_administrators&&@api&&@avatar"
+                tags: "@managing_administrators&&@api"
             javascript: false
